Struct task_master::tm_mod::config::Config

pub struct Config {
    pub name: String,
    pub argv: String,
    pub workingdir: Option<String>,
    pub autostart: bool,
    pub env: Option<Vec<(String, String)>>,
    pub stdout: Option<String>,
    pub stderr: Option<String>,
    pub exitcodes: Vec<i64>,
    pub startretries: u64,
    pub umask: i64,
    pub autorestart: Autorestart,
    pub starttime: Duration,
    pub stopsignal: i64,
    pub stoptime: Duration,
    pub numprocs: i64,
}

Fields
name: String

The Config struct represents all the informations we want to have about a single process we are supervising.

argv: String workingdir: Option<String> autostart: bool env: Option<Vec<(String, String)>> stdout: Option<String> stderr: Option<String> exitcodes: Vec<i64> startretries: u64 umask: i64 autorestart: Autorestart starttime: Duration stopsignal: i64 stoptime: Duration numprocs: i64
Methods

impl Config

pub fn new(
name: String,
argv: String,
workingdir: Option<&str>,
autostart: Option<bool>,
env: Option<Vec<(String, String)>>,
stdout: Option<&str>,
stderr: Option<&str>,
exitcodes: Option<Vec<i64>>,
startretries: Option<i64>,
umask: Option<i64>,
autorestart: Option<&str>,
starttime: Option<i64>,
stopsignal: Option<i64>,
stoptime: Option<i64>,
numprocs: Option<i64>
) -> Config

pub fn from_yaml(name: &str, argv: &str, config: &Yaml) -> Config